+++ aqua_fetch/misc/__init__.py
@@ -1,0 +1,33 @@
+
+
+from ._tabular import Weisssee
+from ._tabular import WaterChemEcuador
+from ._tabular import WaterChemVictoriaLakes
+from ._tabular import WeatherJena
+from ._tabular import WQCantareira
+from ._tabular import WQJordan
+from ._tabular import FlowSamoylov
+from ._tabular import FlowSedDenmark
+from ._tabular import StreamTempSpain
+from ._tabular import RiverTempEroo
+from ._tabular import HoloceneTemp
+from ._tabular import FlowTetRiver
+from ._tabular import SedimentAmersee
+from ._tabular import HydrocarbonsGabes
+from ._tabular import HydroChemJava
+from ._tabular import PrecipBerlin
+from ._tabular import GeoChemMatane
+from ._tabular import WQJordan2
+from ._tabular import YamaguchiClimateJp
+from ._tabular import FlowBenin
+from ._tabular import HydrometricParana
+from ._tabular import RiverTempSpain
+from ._tabular import RiverIsotope
+from ._tabular import EtpPcpSamoylov
+from ._tabular import SWECanada
+from ._tabular import gw_punjab
+from ._tabular import RRAlpineCatchments
+
+from ._tabular import GloHydroRes
+
+from ._hyperspectral import SoilPhosphorus
